Implement #33 - Room DB 테이블 수정
* Constants.kt 수정
  + const val 변수들을 companion object 안에 선언
  + USER_ID 부분은 const val타입이 아니라 var로 선언해서 넣었는데 이렇게 하는게 맞는건지 모르겠음
* tags, pomodoros 테이블에 user id 추가
* Pomodoro의 id 타입을 integer -> long 변경

+++ app/src/main/java/com/yodi/flying/adapters/PomodoroAdapter.kt
@@ -19,7 +19,7 @@ import com.yodi.flying.features.pomolist.PomodoroListFragmentDirections
 import com.yodi.flying.features.pomolist.PomodoroViewModel
 
 
-class PomodoroAdapter(var isTimerRunning: Boolean, var runningPomodoroId: Int) :
+class PomodoroAdapter(var isTimerRunning: Boolean, var runningPomodoroId: Long) :
     ListAdapter<Pomodoro, PomodoroAdapter.ViewHolder>(
         PomodoroDiffCallback()
     ) {
@@ -47,7 +47,7 @@ class PomodoroAdapter(var isTimerRunning: Boolean, var runningPomodoroId: Int) :
     inner class ViewHolder(
         private val binding: ListItemPomodoroBinding,
         private val isTimerRunning: Boolean,
-        private val runningPomodoroId: Int
+        private val runningPomodoroId: Long
     ) : RecyclerView.ViewHolder(binding.root) {
 
 
@@ -72,7 +72,7 @@ class PomodoroAdapter(var isTimerRunning: Boolean, var runningPomodoroId: Int) :
 
         }
         private fun navigateToTimer(
-            pomoId: Int,
+            pomoId: Long,
             view: View
         ) {
             val direction =
@@ -82,7 +82,7 @@ class PomodoroAdapter(var isTimerRunning: Boolean, var runningPomodoroId: Int) :
         }
 
         private fun navigateToPomoDetail(
-            pomoId: Int,
+            pomoId: Long,
             view: View
         ) {
             val direction =
